feat:LottoGame organize
