+++ src/main/java/lotto/Application.java
@@ -2,6 +2,6 @@ package lotto;
 
 public class Application {
     public static void main(String[] args) {
-        // TODO: 프로그램 구현
+
     }
 }
