+++ src/main/java/test/LottoResultTest.java
@@ -2,7 +2,7 @@ package test;
 
 import domain.LottoResult;
 import domain.WinningLotto;
-import lotto.Lotto;
+import domain.Lotto;
 import org.junit.jupiter.api.Test;
 import prize.Prize;
